image_caption/requirements.txt
streamlit 
google-generativeai
Pillow
